Move monster functions to monster.c
