+++ monster.h
@@ -1,0 +1,27 @@
+#ifndef MONSTER_H
+#define MONSTER_H
+
+/**
+ * move_monsters()
+ * Move the monsters
+ */
+void move_monsters(void);
+
+/**
+ * setup_enemy_sprites() - Set up enemy sprite spawn points
+ */
+void setup_enemy_sprites(void);
+
+/**
+ * monster_change_direction()
+ * Change monster direction.
+ */
+void monster_change_direction(void);
+
+/**
+ * move_monsters()
+ * Move the monsters
+ */
+void move_monsters(void);
+
+#endif /* MONSTER_H */
